refactor(windows): move exe selection into Windows class with scoped DI; update callers/tests to match repo patterns
