+++ packages/shorebird_cli/test/src/platform/windows_test.dart
@@ -1,11 +1,21 @@
 import 'dart:io';
 
 import 'package:path/path.dart' as p;
-import 'package:shorebird_cli/src/windows/windows_exe_selector.dart';
+import 'package:scoped_deps/scoped_deps.dart';
+import 'package:shorebird_cli/src/platform/windows.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('selectWindowsAppExe', () {
+  R runWithOverrides<R>(R Function() body) {
+    return runScoped(
+      body,
+      values: {
+        windowsRef.overrideWith(Windows.new),
+      },
+    );
+  }
+
+  group('Windows', () {
     late Directory tempDir;
 
     setUp(() {
@@ -18,7 +28,7 @@ void main() {
 
     test('throws when no .exe files exist', () {
       expect(
-        () => windowsAppExe(tempDir),
+        () => runWithOverrides(() => windows.windowsAppExe(tempDir)),
         throwsA(isA<Exception>()),
       );
     });
@@ -28,7 +38,9 @@ void main() {
         ..createSync();
       final app = File(p.join(tempDir.path, 'myapp.exe'))..createSync();
 
-      final selected = windowsAppExe(tempDir, projectName: 'myapp');
+      final selected = runWithOverrides(
+        () => windows.windowsAppExe(tempDir, projectName: 'myapp'),
+      );
       expect(selected.path, equals(app.path));
 
       // Ensure another executable exists but is not chosen due to projectName.
@@ -39,7 +51,9 @@ void main() {
       final app = File(p.join(tempDir.path, 'cool_game.exe'))..createSync();
       final other = File(p.join(tempDir.path, 'runner.exe'))..createSync();
 
-      final selected = windowsAppExe(tempDir, projectName: 'cool');
+      final selected = runWithOverrides(
+        () => windows.windowsAppExe(tempDir, projectName: 'cool'),
+      );
       expect(selected.path, equals(app.path));
       expect(other.existsSync(), isTrue);
     });
@@ -48,7 +62,7 @@ void main() {
       final a = File(p.join(tempDir.path, 'a.exe'))..createSync();
       final b = File(p.join(tempDir.path, 'b.exe'))..createSync();
 
-      final selected = windowsAppExe(tempDir);
+      final selected = runWithOverrides(() => windows.windowsAppExe(tempDir));
       // Order of listSync is platform dependent but typically creation order.
       // This assertion captures the legacy behavior of returning the first.
       expect(selected.path == a.path || selected.path == b.path, isTrue);
@@ -59,7 +73,10 @@ void main() {
         ..createSync();
       final cp2 = File(p.join(tempDir.path, 'crashpad_wer.exe'))..createSync();
 
-      final selected = windowsAppExe(tempDir);
+      final selected = runScoped(
+        () => windows.windowsAppExe(tempDir),
+        values: {windowsRef.overrideWith(Windows.new)},
+      );
       // When only these executables exist, returns one of them (fallback).
       expect(
         selected.path == cp1.path || selected.path == cp2.path,
